requirements_arxiv.txt
arxiv
google-generativeai
openai
python-dotenv